move gzenv into new package
"gzenv" is the format direnv uses to exchange data structures between
invocations. I'm sure something smarter could be devised but it works
pretty well.
