+++ env_diff.go
@@ -2,6 +2,8 @@ package main
 
 import (
 	"strings"
+
+	"github.com/direnv/direnv/gzenv"
 )
 
 // A list of keys we don't want to deal with
@@ -60,9 +62,9 @@ func BuildEnvDiff(e1, e2 Env) *EnvDiff {
 	return diff
 }
 
-func LoadEnvDiff(base64env string) (diff *EnvDiff, err error) {
+func LoadEnvDiff(gzenvStr string) (diff *EnvDiff, err error) {
 	diff = new(EnvDiff)
-	err = unmarshal(base64env, diff)
+	err = gzenv.Unmarshal(gzenvStr, diff)
 	return
 }
 
@@ -110,7 +112,7 @@ func (self *EnvDiff) Reverse() *EnvDiff {
 }
 
 func (self *EnvDiff) Serialize() string {
-	return marshal(self)
+	return gzenv.Marshal(self)
 }
 
 //// Utils
